Extract create wallet e2e test into own file.
This allows us to run tests separately.
