+++ e2e_tests/src/e2e.test.ts
@@ -11,7 +11,7 @@ const getElementById = async (driver, xpath, timeout = 4000) => {
     return await driver.wait(until.elementIsVisible(el), timeout);
 };
 
-describe("webdriver", () => {
+describe("e2e tests", () => {
     const webAppUrl = "http://localhost:3030";
 
     let driver;
@@ -71,70 +71,8 @@ describe("webdriver", () => {
     }
 
     async function switchToWindow(name: string) {
-        // await driver.switchTo().window(await getWindowHandle(name));
+        await driver.switchTo().window(await getWindowHandle(name));
     }
-
-    test("Create wallet", async () => {
-        const debug = Debug("e2e-create");
-
-        await switchToWindow(extensionTitle);
-
-        debug("Choosing password");
-
-        let step1 = await getElementById(driver, "//button[@data-cy='data-cy-create-wallet-step-1']");
-        await step1.click();
-
-        let mnemonic = "bargain pretty shop spy travel toilet hero ridge critic race weapon elbow";
-
-        let mnemonicInput = await getElementById(driver, "//textarea[@data-cy='data-cy-create-wallet-mnemonic-input']");
-        await mnemonicInput.sendKeys(mnemonic);
-
-        let checkBox = await getElementById(driver, "//label[@data-cy='data-cy-create-wallet-checkbox-input']");
-        await checkBox.click();
-
-        let step2 = await getElementById(driver, "//button[@data-cy='data-cy-create-wallet-step-2']");
-        await step2.click();
-
-        let mnemonicConfirmationInput = await getElementById(
-            driver,
-            "//textarea[@data-cy='data-cy-create-wallet-mnemonic-input-confirmation']",
-        );
-        await mnemonicConfirmationInput.sendKeys(mnemonic);
-
-        let password = "foo";
-        let passwordInput = await getElementById(driver, "//input[@data-cy='data-cy-create-wallet-password-input']");
-        await passwordInput.sendKeys(password);
-
-        debug("Creating wallet");
-        let createWalletButton = await getElementById(
-            driver,
-            "//button[@data-cy='data-cy-create-wallet-button']",
-        );
-        await createWalletButton.click();
-
-        debug("Getting wallet address");
-        let addressField = await getElementById(driver, "//p[@data-cy='data-cy-wallet-address-text-field']");
-        let address = await addressField.getText();
-        debug(`Address found: ${address}`);
-
-        let url = `${webAppUrl}/api/faucet/${address}`;
-        debug("Calling faucet: %s", url);
-        let response = await fetch(url, {
-            method: "POST",
-        });
-        assert(response.ok);
-        let body = await response.text();
-        debug("Faucet response: %s", body);
-
-        // TODO: Remove when automatic balance refreshing is
-        // implemented
-        await new Promise(r => setTimeout(r, 10_000));
-        await driver.navigate().refresh();
-
-        debug("Waiting for balance update");
-        let btcAmount = await getElementById(driver, "//p[@data-cy='data-cy-L-BTC-balance-text-field']", 20_000);
-        debug("Found L-BTC amount: %s", await btcAmount.getText());
-    }, 30_000);
 
     test("sell swap", async () => {
         const debug = Debug("e2e-sell");
